requirements.txt
fastapi~=0.115.12
httpx~=0.28.1
uvicorn~=0.34.1
python-dotenv~=1.1.0